+++ okitweb/static/okit/properties/js/resources/analytics_instance.js
@@ -78,6 +78,18 @@ class AnalyticsInstanceProperties extends OkitResourceProperties {
         const whitelisted_vcns_table = this.createArrayTable('Rules', `${this.id}_whitelisted_vcns`, '', () => this.addWhitelistedVcn())
         this.whitelisted_vcns_tbody = whitelisted_vcns_table.tbody
         this.append(whitelisted_vcns_details.div, whitelisted_vcns_table.table)
+        // Pricing Estimates
+        const pricing_estimates_details = this.createDetailsSection('Pricing Estimates', `${this.id}_pricing_estimates_details`)
+        this.append(this.properties_contents, pricing_estimates_details.details)
+        const pricing_estimates_table = this.createTable('', `${this.id}_pricing_estimates_properties`)
+        this.pricing_estimates_tbody = pricing_estimates_table.tbody
+        this.append(pricing_estimates_details.div, pricing_estimates_table.table)
+        // OCPUs
+        const estimated_ocpu_per_hour_data = {min: 1}
+        const estimated_ocpu_per_hour = this.createInput('number', 'Estimated OCPUs per Hour', `${this.id}_estimated_ocpu_per_hour`, '', (d, i, n) => {n[i].reportValidity(); this.resource.pricing_estimates.estimated_ocpu_per_hour = n[i].value}, estimated_ocpu_per_hour_data)
+        this.append(this.pricing_estimates_tbody, estimated_ocpu_per_hour.row)
+        this.estimated_ocpu_per_hour = estimated_ocpu_per_hour.input
+        this.estimated_ocpu_per_hour_row = estimated_ocpu_per_hour.row
     }
 
     // Load Additional Resource Specific Properties
@@ -101,6 +113,7 @@ class AnalyticsInstanceProperties extends OkitResourceProperties {
         this.vcn_id.property('value', this.resource.network_endpoint_details.vcn_id)
         this.subnet_id.property('value', this.resource.network_endpoint_details.subnet_id)
         this.whitelisted_ips.property('value', this.resource.network_endpoint_details.whitelisted_ips.join(','))
+        this.estimated_ocpu_per_hour.property('value', this.resource.pricing_estimates.estimated_ocpu_per_hour)
         this.loadWhitelistedVcns()
         this.collapseExpandNetworkEndPointInputs()
     }