+++ CHANGELOG.md
@@ -13,12 +13,15 @@
 4. Extend NoSQL Query to display DDL & Indexes in properties.
 5. Drop ExaData from Database System Shapes because this is no longer the preferred implementation.
 6. Implement Version 2 of Cost Estimation / BoM ([Matches new online estimation tool](https://www.oracle.com/uk/cloud/costestimator.html))
+    1. Analytics Instance
     1. Block Storage Volume
     1. File System (Properties allow for usage Estimates)
     1. Instance
+    1. Load Balancer
     1. MySQL Database System
     1. NoSQL Database
     1. Object Storage Bucket (Properties allow for usage Estimates)
+    1. Visual Builder Instance
 ### Bug Fixes
 1. In markdown generation switch to using documentation not definition
 2. Move logging debug message in generate "GET" that accesses unknown element.